feat: add postgresql support

+++ database/migrations/2025_01_04_optimize_plan_table.php
@@ -51,7 +51,7 @@ return new class extends Migration {
                 ->comment('Transfer limit in bytes')->change();
             $table->unsignedInteger('speed_limit')->nullable()
                 ->comment('Speed limit in Mbps, 0 for unlimited')->change();
-            $table->unsignedSmallInteger('reset_traffic_method')->nullable()->default(0)
+            $table->integer('reset_traffic_method')->nullable()->default(0)
                 ->comment('重置流量方式:0跟随系统设置、1每月1号、2按月重置、3不重置、4每年1月1日、5按年重置')->change();
             $table->unsignedInteger('capacity_limit')->nullable()->default(0)
                 ->comment('0 for unlimited')->change();
